feat: basic outline of zarr v3
