--- src/pydantic_zarr/__init__.py
@@ -1,2 +1,0 @@
-# ruff: noqa
-from pydantic_zarr.core import ArraySpec, GroupSpec, to_zarr, from_zarr
