+++ tests/test_v3.py
@@ -1,0 +1,25 @@
+from pydantic_zarr.v3 import ArraySpec, GroupSpec, NamedConfig
+
+
+def test_serialize_deserialize():
+
+    array_attributes = {"foo": 42, "bar": "apples", "baz": [1, 2, 3, 4]}
+
+    group_attributes = {"group": True}
+
+    array_spec = ArraySpec(
+        attributes=array_attributes,
+        shape=[1000, 1000],
+        dimension_names=["rows", "columns"],
+        data_type="float64",
+        chunk_grid=NamedConfig(
+            name="regular", configuration={"chunk_shape": [1000, 100]}
+        ),
+        chunk_key_encoding=NamedConfig(
+            name="default", configuration={"separator": "/"}
+        ),
+        codecs=[NamedConfig(name="GZip", configuration={"level": 1})],
+        fill_value="NaN",
+    )
+
+    GroupSpec(attributes=group_attributes, members={"array": array_spec})
